Scripts/src/databaseupdate/requirements.txt
azure-kusto-data
python-dotenv